LayoutTests/fast/events/platform-wheelevent-in-scrolling-div-expected.txt

PASS event.wheelDeltaY is window.expectedScrollTop*-30
PASS event.wheelDeltaX is window.expectedScrollLeft*-30
PASS event.wheelDelta is window.expectedScrollTop*-30
PASS div.scrollTop is window.expectedScrollTop*window.pixelsPerWheelTick
PASS div.scrollLeft is window.expectedScrollLeft*window.pixelsPerWheelTick
